Split games to own files
and fix bug with account not existing
